cli/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
loguru==0.7.2
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.7
packaging==23.2
pathspec==0.12.1
pinecone-client==2.2.4
platformdirs==4.1.0
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
